Move CTransfer_ConservativeVars
