+++ SU2_CFD/include/transfer/CTransfer.hpp
@@ -504,72 +504,3 @@ public:
 
 };
 
-/*!
- * \class CTransfer_ConservativeVars
- * \brief Transfer conservative variables from a generic zone into another
- * \author R. Sanchez
- * \version 6.2.0 "Falcon"
- */
-
-class CTransfer_ConservativeVars : public CTransfer {
-
-protected:
-
-public:
-
-  /*!
-   * \brief Constructor of the class.
-   */
-  CTransfer_ConservativeVars(void);
-
-  /*!
-   * \overload
-   * \param[in] val_nVar - Number of variables that need to be transferred.
-   * \param[in] config - Definition of the particular problem.
-   */
-  CTransfer_ConservativeVars(unsigned short val_nVar, unsigned short val_nConst, CConfig *config);
-
-  /*!
-   * \brief Destructor of the class.
-   */
-  virtual ~CTransfer_ConservativeVars(void);
-
-  /*!
-   * \brief Retrieve some constants needed for the calculations.
-   * \param[in] donor_solution - Solution from the donor mesh.
-   * \param[in] target_solution - Solution from the target mesh.
-   * \param[in] donor_geometry - Geometry of the donor mesh.
-   * \param[in] target_geometry - Geometry of the target mesh.
-   * \param[in] donor_config - Definition of the problem at the donor mesh.
-   * \param[in] target_config - Definition of the problem at the target mesh.
-   */
-  void GetPhysical_Constants(CSolver *donor_solution, CSolver *target_solution,
-                 CGeometry *donor_geometry, CGeometry *target_geometry,
-                 CConfig *donor_config, CConfig *target_config);
-
-  /*!
-   * \brief Retrieve the variable that will be sent from donor mesh to target mesh.
-   * \param[in] donor_solution - Solution from the donor mesh.
-   * \param[in] donor_geometry - Geometry of the donor mesh.
-   * \param[in] donor_config - Definition of the problem at the donor mesh.
-   * \param[in] Marker_Donor - Index of the donor marker.
-   * \param[in] Vertex_Donor - Index of the donor vertex.
-   * \param[in] Point_Donor - Index of the donor point.
-   */
-  void GetDonor_Variable(CSolver *donor_solution, CGeometry *donor_geometry, CConfig *donor_config,
-               unsigned long Marker_Donor, unsigned long Vertex_Donor, unsigned long Point_Donor);
-
-  /*!
-   * \brief Set the variable that has been received from the target mesh into the target mesh.
-   * \param[in] target_solution - Solution from the target mesh.
-   * \param[in] target_geometry - Geometry of the target mesh.
-   * \param[in] target_config - Definition of the problem at the target mesh.
-   * \param[in] Marker_Target - Index of the target marker.
-   * \param[in] Vertex_Target - Index of the target vertex.
-   * \param[in] Point_Target - Index of the target point.
-   */
-  void SetTarget_Variable(CSolver *target_solution, CGeometry *target_geometry, CConfig *target_config,
-              unsigned long Marker_Target, unsigned long Vertex_Target, unsigned long Point_Target);
-
-
-};
