+++ SU2_CFD/src/transfer/physics/CTransfer_FlowTraction.cpp
@@ -327,43 +327,3 @@ void CTransfer_FlowTraction_DiscAdj::GetPhysical_Constants(CSolver *flow_solutio
 }
 
 
-CTransfer_ConservativeVars::CTransfer_ConservativeVars(void) : CTransfer() {
-
-}
-
-CTransfer_ConservativeVars::CTransfer_ConservativeVars(unsigned short val_nVar, unsigned short val_nConst, CConfig *config) : CTransfer(val_nVar, val_nConst, config) {
-
-}
-
-CTransfer_ConservativeVars::~CTransfer_ConservativeVars(void) {
-
-}
-
-
-void CTransfer_ConservativeVars::GetPhysical_Constants(CSolver *donor_solution, CSolver *target_solution,
-                                                       CGeometry *donor_geometry, CGeometry *target_geometry,
-                                                       CConfig *donor_config, CConfig *target_config) {
-}
-
-void CTransfer_ConservativeVars::GetDonor_Variable(CSolver *donor_solution, CGeometry *donor_geometry, CConfig *donor_config,
-                                                unsigned long Marker_Donor, unsigned long Vertex_Donor, unsigned long Point_Donor) {
-
-  su2double *Solution;
-  unsigned short iVar;
-
-  /*--- Retrieve solution and set it as the donor variable ---*/
-  Solution = donor_solution->node[Point_Donor]->GetSolution();
-
-  for (iVar = 0; iVar < nVar; iVar++)
-    Donor_Variable[iVar] = Solution[iVar];
-}
-
-void CTransfer_ConservativeVars::SetTarget_Variable(CSolver *target_solution, CGeometry *target_geometry,
-                          CConfig *target_config, unsigned long Marker_Target,
-                          unsigned long Vertex_Target, unsigned long Point_Target) {
-
-  /*--- Set the target solution with the value of the Target Variable ---*/
-  target_solution->node[Point_Target]->SetSolution(Target_Variable);
-
-}
-
